Be Stricter on Returned Warps
Rewritten and ended up with nearly the same thing, just additional checks on which connect to what.

Buildings create warps on their own to their parents, and other locations. Check the Building Warps that match the targetName of the origin to retrieve the correct ones, instead of expecting it to be the main entrance.